API testada!

+++ server/app/controllers/meal_controller.py
@@ -19,7 +19,7 @@ router = APIRouter(
 def get_meals(
     current_user: UserRead = Depends(get_current_user), db: Session = Depends(get_db)
 ):
-    return db.query(Meal).filter(Meal.user_id == current_user.id).all()
+    return db.query(Meal).all()
 
 
 @router.get("/{meal_id}", response_model=MealRead)
@@ -67,6 +67,11 @@ def update_meal(
             status_code=status.HTTP_404_NOT_FOUND, detail="Meal not found"
         )
 
+    if meal_db.user_id != current_user.id and not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="You do not have permission to update this meal"
+        )
+
     for key, value in meal.dict(exclude_unset=True).items():
         setattr(meal_db, key, value)
 
@@ -87,6 +92,11 @@ def delete_meal(
     if not meal_db:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Meal not found"
+        )
+
+    if meal_db.user_id != current_user.id and not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="You do not have permission to delete this meal"
         )
 
     db.delete(meal_db)
